requirements.txt
flask
flask-sqlalchemy
pymysql
cryptography
PyJWT
flask-smorest
python-dotenv
marshmallow